% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigershapes.R
\name{batch_read_tigershapes}
\alias{batch_read_tigershapes}
\title{Load geographies from many shapefiles into a tibble of simple features}
\usage{
batch_read_tigershapes(.path_list, .names_to, ...)
}
\arguments{
\item{.path_list}{a list of paths, which may include names}

\item{.names_to}{the output's column that tracks path names}

\item{...}{other arguments for <\link[dplyr:filter]{filtering}> the shapefiles}
}
\value{
a tibble that is also a simple features object.
}
\description{
Load geographies from many shapefiles into a tibble of simple features
}
