% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm.R
\name{central_utm}
\alias{central_utm}
\title{Find the median UTM Zone from a vector of longitudes.}
\usage{
central_utm(.longitudes)
}
\arguments{
\item{.longitudes}{one or more decimal longitudes}
}
\value{
a single integer code for one UTM Zone
}
\description{
Find the median UTM Zone from a vector of longitudes.
}
