% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation_crosswalk.R
\name{crosswalk_spec}
\alias{crosswalk_spec}
\title{Compute the names of the columns in a tabulation crosswalk table}
\usage{
crosswalk_spec(.newer_vintage)
}
\arguments{
\item{.newer_vintage}{\verb{<int>} The more recent of the two decennial censuses being compared.}
}
\value{
\verb{<chr[]>} a named vector of data types for reading the relevant crosswalk's csv file
}
\description{
Tabulation crosswalk tables are used to relate blocks from one decennial
census to another.
They don't always stay the same, especially if the population or land use
in an area has changed rapidly during the ten years between censuses.
}
