% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decennial_pl_variable.R
\name{decennial_pl_variable}
\alias{decennial_pl_variable}
\title{Variable names for PL tables, per the idiosyncrasies of decennial censuses}
\usage{
decennial_pl_variable(.year, .table, .index)
}
\arguments{
\item{.year}{the decade of the census}

\item{.table}{an integer, \link{1, 4}, denoting a PL table.}

\item{.index}{an integer, \link{1, 999}, denoting a variable in a PL table.}
}
\value{
a character vector of variable names.
}
\description{
Decennial censuses record demographic information per Public Law 94-171.
There are four tables (descriptions t.b.d) with the same information each
decade, but the API formatting of the variable names differs for each decade.
This function keeps track of those differences for you.
}
