% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DECENNIAL_POPULATION_VARIABLE_FORMATS}
\alias{DECENNIAL_POPULATION_VARIABLE_FORMATS}
\title{Formats for decennial census variables of overall population.}
\format{
\subsection{\code{DECENNIAL_POPULATION_VARIABLE_FORMATS} A data frame with 3 rows}{

and 4 columns:
\describe{
\item{Year}{Decadal years, e.g. 2010, 2020}
\item{Group}{The table of data that the variables are found in.}
\item{Separator}{The character, if any, that goes between the group name and the variable number}
\item{Suffix}{The suffix, if any, that goes on the end of a variable name}
}
}
}
\source{
\url{https://api.census.gov/data/2000/dec/pl/groups.html}

\url{https://api.census.gov/data/2010/dec/pl/groups.html}

\url{https://api.census.gov/data/2020/dec/pl/groups.html}
}
\usage{
DECENNIAL_POPULATION_VARIABLE_FORMATS
}
\description{
A tibble with formatting data about the population variable for each
decennial census
}
\keyword{datasets}
