% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DECENNIAL_TABULATION_CROSSWALK}
\alias{DECENNIAL_TABULATION_CROSSWALK}
\title{The layout for a crosswalk of tabulation blocks between decennial censuses}
\format{
\subsection{\code{DECENNIAL_TABULATION_CROSSWALK}}{

A data frame with 18 rows and 5 columns.
\describe{
\item{Role}{\verb{<chr>} "FIPS" for parts of GEOIDs, "Area" for block size, and "Part" for information about overlaps}
\item{Detail}{\verb{<chr>} several values are possible, depending upon the \code{Role}}
\item{Data Type}{\verb{<chr>} usually "c," but occasionally "n" for areas and overlaps.}
\item{Pre 2000}{\verb{<lgl>} Does this column appear in earlier crosswalks?}
\item{Field Description}{\verb{<chr>} a short but informative definition of the field's role.}
}
}
}
\source{
\url{https://www.census.gov/programs-surveys/geography/technical-documentation/records-layout/2020-census-block-record-layout.html}

\url{https://www.census.gov/programs-surveys/geography/technical-documentation/records-layout/2000-relationship-file-record-layout.html}
}
\usage{
DECENNIAL_TABULATION_CROSSWALK
}
\description{
Blocks from one decennial census do not necessarily correspond to those from
others.
A crosswalk between tabulation blocks is used to define those correspondences.
}
\seealso{
\code{\link[=crosswalk_field_name]{crosswalk_field_name()}}
}
\keyword{datasets}
