% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_latlong_to_numeric.R
\name{fix_latlong_to_numeric}
\alias{fix_latlong_to_numeric}
\title{Convert interpreted latitude and longitude from text to numbers}
\usage{
fix_latlong_to_numeric(.features)
}
\arguments{
\item{.features}{a tibble with fields that start with "INTPT"}
}
\value{
the same tibble, but with those fields converted to numbers
}
\description{
The census fields of "INTPTLAT" and "INTPTLON" contain "+" and "-" characters.
For this reason, \code{sf::read_sf} interprets them as character, not numeric, columns.
This function rectifies that.
}
