% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names_by_coalescing.R
\name{fix_names_by_coalescing}
\alias{fix_names_by_coalescing}
\title{Coalesce NAME and NAMELSAD fields to fix blank NAME entries}
\usage{
fix_names_by_coalescing(.features)
}
\arguments{
\item{.features}{any tibble with NAME and NAMELSAD features.}
}
\value{
a tibble with an additional column, "Name," and neither "NAME" nor
"NAMELSAD"
}
\description{
Census TIGER/Line shapefiles have a NAME and a NAMELSAD field, but the NAME
one is sometimes blank. The NAMELSAD values are often more informative. That
is especially true for things like blocks and tracts. This function condenses
both into NAME but doesn't remove NAMELSAD. It prefers the values from
NAMELSAD, if they are different. Both NAME and NAMELSAD are removed.
}
\examples{
data.frame(
  NAME = c("Townsville", "", NA),
  NAMELSAD = c("The City of Townsville", "Metrograd", "Locationopolis")
) |> fix_names_by_coalescing()
}
