% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoid_to_tract_id.R
\name{geoid_to_tract_id}
\alias{geoid_to_tract_id}
\title{Extract the 4- or 6-digit tract codes from full census GEOID}
\usage{
geoid_to_tract_id(.geoids, .years)
}
\arguments{
\item{.geoids}{a vector of GEOIDs for blocks, block groups, or tracts.}

\item{.years}{the census vintage year, as an integer, for each GEOID}
}
\value{
a character vector of zero-padded 4- or 6-digit tract codes
}
\description{
Extract the 4- or 6-digit tract codes from full census GEOID
}
