% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{geom_terrain}
\alias{geom_terrain}
\title{Plot census blocks by water or land use}
\usage{
geom_terrain(.blocks, .line_color = "#cccccc66", .scale_name = NULL)
}
\arguments{
\item{.blocks}{a simple features table from the Census TIGER/Line database}

\item{.line_color}{optional, the color for block borders, defaults to light gray}

\item{.scale_name}{optional, a title for the fill scale, defaults to empty}
}
\value{
an list of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} commands, suitable for chaining
}
\description{
Plot census blocks by water or land use
}
