% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercgeos-package.R
\docType{package}
\name{hercgeos-package}
\alias{hercgeos}
\alias{hercgeos-package}
\title{hercgeos: Analyze and Manipulate Spatial and Geometric Data from the US Census}
\description{
Utilities for wrangling field names, geometry boundaries, and coordinate systems for shapefiles from the US Census Bureau's TIGER/Line database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hercgeos/}
  \item \url{https://github.com/higherX4Racine/hercgeos}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
