% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{is_water}
\alias{is_water}
\title{Identify geographies that are all water}
\usage{
is_water(.water_area, .land_area)
}
\arguments{
\item{.water_area}{\if{html}{\out{<numeric>}} the area of a geography that is water}

\item{.land_area}{\if{html}{\out{<numeric>}} the area of a geography that is land}
}
\value{
a logical vector that is \code{TRUE} for water areas
}
\description{
Identify geographies that are all water
}
