% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NASHVILLE_NH}
\alias{NASHVILLE_NH}
\title{Twelve 2020 Census tracts from the northern part of Nashua, NH.}
\format{
\subsection{\code{NASHVILLE_NH} A list of three tibbles of simple features}{

\strong{blocks} : All census blocks with GEOIDs that start with "33011010"

\strong{tracts} : All census tracts with GEOIDs that start with "33011010"

\strong{historic_district} : The registered national historic place that once was
Nashville, NH.

For \code{blocks} and \code{tracts}, the columns are verbatim what you find in
freshly-downloaded TIGER/Line shapefiles. For \code{historic_district}, the
columns are "RESNAME", "SRC_DAT", and "NR_PROP"
}
}
\source{
The National Register of Historic Places' database of shapefiles:
\url{https://irma.nps.gov/DataStore/Reference/Profile/2210280}

The US Censu Bureau's TIGER/Line shapefile database:
\url{https://www.census.gov/cgi-bin/geo/shapefiles/index.php}
}
\usage{
NASHVILLE_NH
}
\description{
This set of simple features are all of the 2020 Census tracts whose GEOIDs
begin with "33011010". That means that they are in Hillsborough County, NH.
It was created by filtering, but otherwise not changing, a file with every
2020 Census tract in New Hampshire.
}
\details{
The northern part of the city of Nashua, NH once considered seceding and
becoming its own city, Nashville. These tracts are roughly in that area.
}
\keyword{datasets}
