% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation_crosswalk.R
\name{perform_crosswalk}
\alias{perform_crosswalk}
\title{Join the blocks from consecutive decennial censuses by their crosswalks}
\usage{
perform_crosswalk(.blocks, .crosswalk, .vintage)
}
\arguments{
\item{.blocks}{\verb{<tbl>} a data frame with columns "State," "County," "Tract," and "Block."]}

\item{.crosswalk}{\verb{<tbl>} a product of \code{\link[=read_crosswalk]{read_crosswalk()}}, from 10 years earlier to \code{.vintage}}

\item{.vintage}{\verb{<int>} the decennial year that defined \code{.blocks}}
}
\value{
\verb{<tbl>} a lookup table from later blocks to earlier ones
\describe{
\item{Vintage}{\verb{<int>} the later decennial census}
\item{GEOID}{\verb{<chr>} identifiers for blocks in the later decennial census}
\item{State}{\verb{<chr>} the state for a block in the earlier census.}
\item{County}{\verb{<chr>} the county for a block in the earlier census.}
\item{Tract}{\verb{<chr>} the tract for a block in the earlier census.}
\item{Block}{\verb{<chr>} the id for a block in the earlier census.}
\item{Overlap}{\verb{<dbl>} the proportion of the older block that intersects with the newer one}
}
}
\description{
Join the blocks from consecutive decennial censuses by their crosswalks
}
