% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation_crosswalk.R
\name{read_crosswalk}
\alias{read_crosswalk}
\title{Read a file with crosswalk information between the blocks of two consecutive decennial censuses.}
\usage{
read_crosswalk(.filepath, .newer_vintage, ...)
}
\arguments{
\item{.filepath}{\verb{<chr>} the full path to the crosswalk file, which may be a ZIP archive}

\item{.newer_vintage}{\verb{<int>} The more recent of the two decennial censuses being compared.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Additional arguments for \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\value{
\verb{<tbl>} a many-to-many relationship table between the two census's blocks.
}
\description{
Read a file with crosswalk information between the blocks of two consecutive decennial censuses.
}
\seealso{
\code{\link[=crosswalk_spec]{crosswalk_spec()}}
}
