% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigershapes.R
\name{read_tigershapes}
\alias{read_tigershapes}
\title{Import data from US Census TIGER/Line shapefiles}
\usage{
read_tigershapes(.path)
}
\arguments{
\item{.path}{string, the location of the shapefile directory}
}
\value{
a simple features table with one entry per geography
}
\description{
Import data from US Census TIGER/Line shapefiles
}
\examples{
library(sf)

}
