% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TERRAIN_COLORS}
\alias{TERRAIN_COLORS}
\title{A named character vector of hex-encoded colors for water, developed and undeveloped land.}
\format{
An object of class \code{character} of length 3.
}
\usage{
TERRAIN_COLORS
}
\description{
A named character vector of hex-encoded colors for water, developed and undeveloped land.
}
\keyword{datasets}
