% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terrain_factor}
\alias{terrain_factor}
\title{Create a factor of terrains based on census reporting of water area and land
use.}
\usage{
terrain_factor(.use_type, ...)
}
\arguments{
\item{.use_type}{character The census labels blocks as "U" for developed and
"R" for undeveloped}

\item{...}{
  Arguments passed on to \code{\link[=is_water]{is_water}}
  \describe{
    \item{\code{.water_area}}{\if{html}{\out{<numeric>}} the area of a geography that is water}
    \item{\code{.land_area}}{\if{html}{\out{<numeric>}} the area of a geography that is land}
  }}
}
\value{
a factor with three levels, the names of \link{TERRAIN_COLORS}
}
\description{
Create a factor of terrains based on census reporting of water area and land
use.
}
\seealso{
\code{\link[=is_water]{is_water()}}
}
