% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TIGER_ATTRIBUTES}
\alias{TIGER_ATTRIBUTES}
\title{A named character vector of shapefile attributes for data from TIGER/Line}
\format{
An object of class \code{character} of length 25.
}
\usage{
TIGER_ATTRIBUTES
}
\description{
A named character vector of shapefile attributes for data from TIGER/Line
}
\keyword{datasets}
