% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_name_for.R
\name{tiger_name_for}
\alias{tiger_name_for}
\title{Create a tigershape filename based upon state, year, and geography}
\usage{
tiger_name_for(
  state,
  year,
  geography = c("county subunit", "school district", "place", "tract", "block group",
    "block")
)
}
\arguments{
\item{state}{\if{html}{\out{<int>}} the two-digit FIPS number for the state}

\item{year}{\if{html}{\out{<int>}} the four-digit census year}

\item{geography}{<chr, optional> one of "county subunits" (the default), "school districts", "places", "tracts", "blocks"}
}
\value{
a string such as "tl_2000_55_cousub00"
}
\description{
Create a tigershape filename based upon state, year, and geography
}
\examples{
tiger_name_for(55L, 2000L, "county subunit")
tiger_name_for(25L, 2010L, "school district")
tiger_name_for(10L, 2020L, "place")
tiger_name_for(2L, 2020L, "block")
}
