% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm.R
\name{transform_tiger_to_utm}
\alias{transform_tiger_to_utm}
\title{Transform native TIGER/Line lat-long features to UTM features}
\usage{
transform_tiger_to_utm(
  .tiger_features,
  longitude_label = "INTPTLON",
  datum = "NAD83"
)
}
\arguments{
\item{.tiger_features}{a shapefile from the Census' TIGER/Line database}

\item{longitude_label}{optional, the column with longitudes, defaults to "INTPTLON"}

\item{datum}{optional, the datum that the shapefile uses, defaults to "NAD83}
}
\value{
a <\link[sf:sf]{sf::sf}> object
}
\description{
Transform native TIGER/Line lat-long features to UTM features
}
\seealso{
\code{\link[=central_utm]{central_utm()}}

\code{\link[=utm_crs]{utm_crs()}}
}
