% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_decennial_tract_codes.R
\name{truncate_decennial_tract_codes}
\alias{truncate_decennial_tract_codes}
\title{Lop off trailing "00"s from census tract codes from 2000.}
\usage{
truncate_decennial_tract_codes(.tract_codes, .years)
}
\arguments{
\item{.tract_codes}{a character vector of tract labels like "999999" or "000100"}

\item{.years}{numeric years.}
}
\value{
a character vector of tract labels like "999999" or "0001"
}
\description{
Lop off trailing "00"s from census tract codes from 2000.
}
