% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNIT_PENTAGON}
\alias{UNIT_PENTAGON}
\title{An \code{sfc} object that is a five-sided polygon with area 1.}
\format{
An object of class \code{sfc_POLYGON} (inherits from \code{sfc}) of length 1.
}
\usage{
UNIT_PENTAGON
}
\description{
An \code{sfc} object that is a five-sided polygon with area 1.
}
\keyword{datasets}
