% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm.R
\name{utm_crs}
\alias{utm_crs}
\title{Create a coordinate reference system for a UTM Zone and datum.}
\usage{
utm_crs(utm_zone, datum = "NAD83")
}
\arguments{
\item{utm_zone}{an integer that refers to a single zone}

\item{datum}{the name of a model of the earth's shape, such as "NAD83"}
}
\value{
an <\code{\link[sf:st_crs]{sf::st_crs()}}> object
}
\description{
Create a coordinate reference system for a UTM Zone and datum.
}
