% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm.R
\name{utm_zone_from_longitude}
\alias{utm_zone_from_longitude}
\title{Identify the UTM Zone that corresponds to a longitude}
\usage{
utm_zone_from_longitude(.longitude)
}
\arguments{
\item{.longitude}{a decimal longitude}
}
\value{
an integer UTM code
}
\description{
Identify the UTM Zone that corresponds to a longitude
}
