% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_old_geoids.R
\name{fix_old_geoids}
\alias{fix_old_geoids}
\title{Replace ID fields with "GEOID"}
\usage{
fix_old_geoids(.features)
}
\arguments{
\item{.features}{a data frame}
}
\value{
the same data frame, but with one column possibly renamed.
}
\description{
Older (pre-2010) shapefiles have geography-specific names for their GEOID
fields. For example, a county subdivision shapefile would have a field named
"COSBIDFP." This function renames any field that includes "IDFP" to "GEOID."
}
\examples{
data.frame(
  COSIBIDFP = c("55100", "55101", "551012")
) |>
fix_old_geoids()

}
