% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_filter.R
\name{geometry_filter}
\alias{geometry_filter}
\title{Filter one or more geometries by a reference geometry using a single geometric predicate.}
\usage{
geometry_filter(.target_geoms, .reference_geom, .predicate, ...)
}
\arguments{
\item{.target_geoms}{A table of one or more to-be-filtered simple features.}

\item{.reference_geom}{A single simple feature that the targets will be compared to.}

\item{.predicate}{A <\link[sf:geos_binary_pred]{binary geometric predicate}>}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments for \link[tidyselect:language]{selecting} columns from \code{.target_geoms}}
}
\value{
a subset of \code{.target_geoms} that meets the criteria represented by \code{.reference_geom} and \code{.predicate}
}
\description{
Filter one or more geometries by a reference geometry using a single geometric predicate.
}
