% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{is_water}
\alias{is_water}
\title{Identify census blocks that are all water}
\usage{
is_water(.features, .water_area_label = "AWATER", .land_area_label = "ALAND")
}
\arguments{
\item{.features}{a simple features table}

\item{.water_area_label}{optional, the column that describes water area,
defaults to "AWATER"}

\item{.land_area_label}{optional, the column that describes land area,
defaults to "ALAND"}
}
\value{
a logical vector that is \code{TRUE} for water areas
}
\description{
Identify census blocks that are all water
}
