% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terrain_factor}
\alias{terrain_factor}
\title{Create a factor of terrains based on census reporting of water area and land use.}
\usage{
terrain_factor(.blocks, ...)
}
\arguments{
\item{.blocks}{a simple features table from TIGER/Line}
}
\value{
a factor with three levels, the names of \link{TERRAIN_COLORS}
}
\description{
Create a factor of terrains based on census reporting of water area and land use.
}
