% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_decennial_tract_codes.R
\name{truncate_decennial_tract_codes}
\alias{truncate_decennial_tract_codes}
\title{Lop off trailing "00"s from census tract codes from 2000.}
\usage{
truncate_decennial_tract_codes(.tract_codes, .years)
}
\arguments{
\item{.tract_codes}{a character vector of labels like "999999" or "000100"}

\item{.years}{numeric years.}
}
\value{
a character vector of labels like "999999" or "0001" (or "000100").
}
\description{
For inscrutable reasons, the census requires 4-digit tract numbers in some
API calls, but ONLY for tracts from the year 2000 that end in "00"
}
