% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_pentagon}
\alias{unit_pentagon}
\title{An \code{sfc} object with one five-sided polygon whose are is one.}
\format{
An object of class \code{sfc_POLYGON} (inherits from \code{sfc}) of length 1.
}
\usage{
unit_pentagon
}
\description{
An \code{sfc} object with one five-sided polygon whose are is one.
}
\keyword{datasets}
